+++ src/leveldb/port/port_posix.h
@@ -109,7 +109,7 @@ class CondVar {
   explicit CondVar(Mutex* mu);
   ~CondVar();
   void Wait();
-  bool Wait(int32_t wait_millisec);
+  bool Wait(int64_t wait_millisec);
   void Signal();
   void SignalAll();
  private:
